+++ src/com/szczypiorofix/sweetrolls/game/objects/characters/OtherCharacter.java
@@ -11,17 +11,17 @@ import java.util.ArrayList;
 public class OtherCharacter extends Character {
 
 
-    OtherCharacter(String name, int x, int y, int width, int height, ObjectType objectType, ArrayList<Property> properties) {
+    OtherCharacter(String name, float x, float y, float width, float height, ObjectType objectType, ArrayList<Property> properties) {
         super(name, x, y, width, height, objectType, properties);
     }
 
     @Override
-    public void update(GameContainer gc, StateBasedGame sbg, int delta, int offsetX, int offsetY) {
+    public void update(GameContainer gc, StateBasedGame sbg, int delta, float offsetX, float offsetY) {
 
     }
 
     @Override
-    public void render(GameContainer gc, StateBasedGame sbg, Graphics g, int offsetX, int offsetY) {
+    public void render(GameContainer gc, StateBasedGame sbg, Graphics g, float offsetX, float offsetY) {
 
     }
