+++ src/com/szczypiorofix/sweetrolls/game/objects/item/Place.java
@@ -23,12 +23,12 @@ public class Place extends Item {
     }
 
     @Override
-    public void update(GameContainer gc, StateBasedGame sbg, int delta, int offsetX, int offsetY) throws SlickException {
+    public void update(GameContainer gc, StateBasedGame sbg, int delta, float offsetX, float offsetY) throws SlickException {
         hover = false;
     }
 
     @Override
-    public void render(GameContainer gc, StateBasedGame sbg, Graphics g, int offsetX, int offsetY) throws SlickException {
+    public void render(GameContainer gc, StateBasedGame sbg, Graphics g, float offsetX, float offsetY) throws SlickException {
         image.draw(- offsetX + x, - offsetY + y);
         if (hover) {
             font.draw(name, - offsetX + x, - offsetY + y - 20);
